LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/snap-to-visible-areas-y-axis-expected.txt


FAIL Only snap to visible area on Y axis, even when the non-visible ones are closer assert_equals: expected 800 but got 500
